% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.forms.R
\name{amber.forms}
\alias{amber.forms}
\title{Get the forms}
\usage{
amber.forms(
  amber,
  study = NULL,
  query = list(),
  from = 0,
  limit = 100,
  df = TRUE
)
}
\arguments{
\item{amber}{A Amber object}

\item{study}{Study name or identifier, optional}

\item{query}{The search query}

\item{from}{From item}

\item{limit}{Max number of items}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the forms
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.forms(a)
amber.forms(a, study="Trauma Registry", from=0, limit=10)
amber.logout(a)
}
}
\seealso{
Other studies functions: 
\code{\link{amber.case_report_export}()},
\code{\link{amber.case_report_forms}()},
\code{\link{amber.case_report_form}()},
\code{\link{amber.form_revisions}()},
\code{\link{amber.form_revision}()},
\code{\link{amber.form}()},
\code{\link{amber.studies}()},
\code{\link{amber.study}()}
}
\concept{studies functions}
