% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.users.R
\name{amber.group}
\alias{amber.group}
\title{Get a group}
\usage{
amber.group(amber, id)
}
\arguments{
\item{amber}{An Amber object}

\item{id}{Group's name or identifier}
}
\value{
A group object as a named list
}
\description{
Get a user by name or identifier.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.group(a, id = "obiba.org")
amber.group(a, id = "615abb061799dd6a2cdbbb10")
amber.logout(a)
}
}
\seealso{
Other users functions: 
\code{\link{amber.groups}()},
\code{\link{amber.user}()},
\code{\link{amber.users}()}
}
\concept{users functions}
