% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.interviews.R
\name{amber.interviews_steps}
\alias{amber.interviews_steps}
\title{Get the steps of interview records}
\usage{
amber.interviews_steps(interviews)
}
\arguments{
\item{interviews}{The data frame of interviews}
}
\value{
A data.frame of steps
}
\description{
Get the steps of the provided interviews.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")

# Find interview records of a study
itws <- amber.interviews(a, study = "Trauma Registry")

# Get steps
steps <- amber.interviews_steps(itws)

amber.logout(a)
}
}
\seealso{
Other studies functions: 
\code{\link{amber.campaign}()},
\code{\link{amber.campaigns}()},
\code{\link{amber.case_report_export}()},
\code{\link{amber.case_report_form}()},
\code{\link{amber.case_report_forms}()},
\code{\link{amber.case_reports}()},
\code{\link{amber.form}()},
\code{\link{amber.form_revision}()},
\code{\link{amber.form_revisions}()},
\code{\link{amber.forms}()},
\code{\link{amber.interview_design}()},
\code{\link{amber.interview_designs}()},
\code{\link{amber.interview_export}()},
\code{\link{amber.interviews}()},
\code{\link{amber.interviews_step_actions}()},
\code{\link{amber.interviews_step_data}()},
\code{\link{amber.participant}()},
\code{\link{amber.participants}()},
\code{\link{amber.studies}()},
\code{\link{amber.study}()}
}
\concept{studies functions}
