% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.R
\name{amber.logout}
\alias{amber.logout}
\title{Close connection with Amber}
\usage{
amber.logout(amber)
}
\arguments{
\item{amber}{An Amber object}
}
\value{
No value
}
\description{
Close connection and release resources of Amber.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.logout(a)
}
}
