% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.users.R
\name{amber.subjects}
\alias{amber.subjects}
\title{Get subjects}
\usage{
amber.subjects(amber, query = list(), df = TRUE)
}
\arguments{
\item{amber}{An Amber object}

\item{query}{The search query}

\item{df}{Return a data.frame (default is TRUE)}
}
\value{
A data.frame (or a list of raw results when 'df' is FALSE)
}
\description{
Get the users and groups as subjects: an identifier, a type and a name.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.subjects(a)
amber.logout(a)
}
}
