% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.tasks.R
\name{amber.task}
\alias{amber.task}
\title{Get a task}
\usage{
amber.task(amber, id)
}
\arguments{
\item{amber}{An Amber object}

\item{id}{Task's identifier}
}
\value{
An task object as a named list
}
\description{
Get a task by its identifier.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.task(a, id = "61e69a22fea2df2f3108b508")
amber.logout(a)
}
}
\seealso{
Other tasks functions: 
\code{\link{amber.task_logs}()},
\code{\link{amber.tasks}()},
\code{\link{amber.tasks_logs}()}
}
\concept{tasks functions}
