% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.tasks.R
\name{amber.task_logs}
\alias{amber.task_logs}
\title{Get the logs of a task}
\usage{
amber.task_logs(amber, id, df = TRUE)
}
\arguments{
\item{amber}{An Amber object}

\item{id}{Task's identifier}

\item{df}{Return a data.frame (default is TRUE)}
}
\value{
A data.frame (or a named list of raw results when 'df' is FALSE)
}
\description{
Get the logs of a task by its identifier.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.task(a, id = "61e69a22fea2df2f3108b508")
amber.logout(a)
}
}
\seealso{
Other tasks functions: 
\code{\link{amber.task}()},
\code{\link{amber.tasks}()},
\code{\link{amber.tasks_logs}()}
}
\concept{tasks functions}
