% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.tasks.R
\name{amber.tasks_logs}
\alias{amber.tasks_logs}
\title{Get the tasks logs}
\usage{
amber.tasks_logs(amber, query = list(), skip = 0, limit = 100, df = TRUE)
}
\arguments{
\item{amber}{An Amber object}

\item{query}{The search query}

\item{skip}{Number of items to skip}

\item{limit}{Max number of items}

\item{df}{Return a data.frame (default is TRUE)}
}
\value{
A data.frame (or a named list of raw results when 'df' is FALSE)
}
\description{
Get the tasks logs
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.tasks_logs(a, skip = 0, limit = 10)
amber.logout(a)
}
}
\seealso{
Other tasks functions: 
\code{\link{amber.task}()},
\code{\link{amber.task_logs}()},
\code{\link{amber.tasks}()}
}
\concept{tasks functions}
