% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.R
\name{.delete}
\alias{.delete}
\title{Issues a DELETE request to Amber for the specified resource}
\usage{
.delete(amber, ..., query = list())
}
\description{
Issues a DELETE request to Amber for the specified resource
}
\keyword{internal}
