% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.R
\name{.get}
\alias{.get}
\title{Issues a GET request to Amber for the specified resource}
\usage{
.get(amber, ..., query = list())
}
\description{
Issues a GET request to Amber for the specified resource
}
\keyword{internal}
