% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.makeDictionary}
\alias{.makeDictionary}
\title{Turns a dictionary in JSON format into a "Opal" one (variables and categories)}
\usage{
.makeDictionary(vars)
}
\description{
Turns a dictionary in JSON format into a "Opal" one (variables and categories)
}
\keyword{internal}
