% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.R
\name{.post}
\alias{.post}
\title{Issues a POST form request to Amber for the specified resource}
\usage{
.post(amber, ..., query = list())
}
\description{
Issues a POST form request to Amber for the specified resource
}
\keyword{internal}
