% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.R
\name{.url}
\alias{.url}
\title{Utility method to build urls. Concatenates all arguments and adds a '/' separator between each element}
\usage{
.url(amber, ...)
}
\description{
Utility method to build urls. Concatenates all arguments and adds a '/' separator between each element
}
\keyword{internal}
