% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.case-reports.R
\name{amber.case_report_export}
\alias{amber.case_report_export}
\title{Get the case report forms}
\usage{
amber.case_report_export(
  amber,
  study = NULL,
  form = NULL,
  query = list(),
  from = 0,
  limit = 100,
  df = FALSE
)
}
\arguments{
\item{amber}{A Amber object}

\item{study}{Study identifier (name or id), optional.}

\item{form}{Form identifier (name or id), optional.}

\item{query}{The search query}

\item{from}{From item}

\item{limit}{Max number of items}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the case reports of one or several form(s).
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.case_report_forms(a, form="61e69a22fea2df2f3108b508", from=0, limit=10)
amber.case_report_forms(a, form="Adult trauma")
amber.case_report_forms(a, study="Trauma Registry", query = list(revision = 1))
amber.case_report_forms(a, query = list(revision = 1))
amber.logout(a)
}
}
\seealso{
Other studies functions: 
\code{\link{amber.case_report_forms}()},
\code{\link{amber.case_report_form}()},
\code{\link{amber.form_revisions}()},
\code{\link{amber.form_revision}()},
\code{\link{amber.forms}()},
\code{\link{amber.form}()},
\code{\link{amber.studies}()},
\code{\link{amber.study}()}
}
\concept{studies functions}
