% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.case-reports.R
\name{amber.case_report_export}
\alias{amber.case_report_export}
\title{Get the case report records}
\usage{
amber.case_report_export(
  amber,
  study = NULL,
  form = NULL,
  from = NULL,
  to = NULL,
  pId = NULL,
  query = list(),
  skip = 0,
  limit = 100,
  df = TRUE
)
}
\arguments{
\item{amber}{A Amber object}

\item{study}{Study identifier (name or id), optional.}

\item{form}{Form identifier (name or id), optional.}

\item{from}{From date (included), optional}

\item{to}{To date (included), optional}

\item{pId}{Patient/participant identifier}

\item{query}{The search query}

\item{skip}{Number of items to skip}

\item{limit}{Max number of items}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the case report records of one or several form(s).
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")

# Find all case reports
amber.case_report_export(a)

# Find all case reports in a range of time
amber.case_report_export(a, from = "2022-01-12 00:00", to = "2022-02-13")

# Find all case reports for a specific participant/patient identifier
amber.case_report_export(a, pId = "1231")

# Find all case reports having their identifier matching a regular expression
amber.case_report_export(a, query = list(`data._id[$search]` = "^12"))

# Export records collected with a study's form in a specific version
amber.case_report_export(a,
  study = "Trauma Registry",
  form = "Adult trauma",
  query = list(revision = 6))

# Export records collected with a study's form in all versions used
tables <- amber.case_report_export(a,
  study = "Trauma Registry",
  form = "Adult trauma")

# Result contains both data and dictionary
tables

# Tables are named with the <form name>-<revision> pattern
names(tables)

# Merge datasets from different versions if relevant
dplyr::bind_rows(lapply(tables, function (t) {
  t$data
}))

amber.logout(a)
}
}
\seealso{
Other studies functions: 
\code{\link{amber.case_report_forms}()},
\code{\link{amber.case_report_form}()},
\code{\link{amber.form_revisions}()},
\code{\link{amber.form_revision}()},
\code{\link{amber.forms}()},
\code{\link{amber.form}()},
\code{\link{amber.studies}()},
\code{\link{amber.study}()}
}
\concept{studies functions}
