% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.case-reports.R
\name{amber.case_report_form}
\alias{amber.case_report_form}
\title{Get a form}
\usage{
amber.case_report_form(amber, form, revision = NULL)
}
\arguments{
\item{amber}{A Amber object}

\item{form}{Form's name or identifier}

\item{revision}{Revision number, optional, default is NULL (means that the case report form uses the latest form revision).}
}
\description{
Get a case report form by its form name or identifier and its revision number (if any).
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.case_report_form(a, form = "Adult trauma", revision = 10)
amber.case_report_form(a, form = "61e69a22fea2df2f3108b508")
amber.logout(a)
}
}
\seealso{
Other studies functions: 
\code{\link{amber.case_report_export}()},
\code{\link{amber.case_report_forms}()},
\code{\link{amber.form_revisions}()},
\code{\link{amber.form_revision}()},
\code{\link{amber.forms}()},
\code{\link{amber.form}()},
\code{\link{amber.studies}()},
\code{\link{amber.study}()}
}
\concept{studies functions}
