% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.studies.R
\name{amber.studies}
\alias{amber.studies}
\title{Get the studies}
\usage{
amber.studies(amber, query = list(), skip = 0, limit = 100, df = TRUE)
}
\arguments{
\item{amber}{A Amber object}

\item{query}{The search query}

\item{skip}{Number of items to skip}

\item{limit}{Max number of items}

\item{df}{Return a data.frame (default is TRUE)}
}
\description{
Get the studies
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.studies(a, skip=0, limit=10)
amber.logout(a)
}
}
\seealso{
Other studies functions: 
\code{\link{amber.case_report_export}()},
\code{\link{amber.case_report_forms}()},
\code{\link{amber.case_report_form}()},
\code{\link{amber.form_revisions}()},
\code{\link{amber.form_revision}()},
\code{\link{amber.forms}()},
\code{\link{amber.form}()},
\code{\link{amber.study}()}
}
\concept{studies functions}
