% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.users.R
\name{amber.user}
\alias{amber.user}
\title{Get a user}
\usage{
amber.user(amber, id)
}
\arguments{
\item{amber}{A Amber object}

\item{id}{User's email or identifier}
}
\description{
Get a user by email or identifier.
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.user(a, id = "jim@muse.com")
amber.user(a, id = "6151b512268f582926d37f90")
amber.logout(a)
}
}
\seealso{
Other users functions: 
\code{\link{amber.groups}()},
\code{\link{amber.group}()},
\code{\link{amber.users}()}
}
\concept{users functions}
