% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amber.users.R
\name{amber.users}
\alias{amber.users}
\title{Get the users}
\usage{
amber.users(amber, query = list(), skip = 0, limit = 100, df = TRUE)
}
\arguments{
\item{amber}{An Amber object}

\item{query}{The search query}

\item{skip}{Number of items to skip}

\item{limit}{Max number of items}

\item{df}{Return a data.frame (default is TRUE)}
}
\value{
A data.frame (or a named list of raw results when 'df' is FALSE)
}
\description{
Get the users
}
\examples{
\dontrun{
a <- amber.login("https://amber-demo.obiba.org")
amber.users(a, skip = 0, limit = 10)
amber.logout(a)
}
}
\seealso{
Other users functions: 
\code{\link{amber.group}()},
\code{\link{amber.groups}()},
\code{\link{amber.user}()}
}
\concept{users functions}
